requirements.txt
Django
opencv-python
scikit-learn
pymysql
Pillow
matplotlib
opencv-contrib-python